Rename verifier to abiProvider
