+++ apps/contract-interaction/src/app/abiProviders/AbstractAbiProvider.ts
@@ -2,13 +2,13 @@ import { CompilerAbstract } from '@remix-project/remix-solidity'
 import type { LookupResponse, SubmittedContract, VerificationResponse } from '../types'
 
 // Optional function definitions
-export interface AbstractVerifier {
+export interface AbstractAbiProvider {
   verifyProxy(submittedContract: SubmittedContract): Promise<VerificationResponse>
   checkVerificationStatus?(receiptId: string): Promise<VerificationResponse>
   checkProxyVerificationStatus?(receiptId: string): Promise<VerificationResponse>
 }
 
-export abstract class AbstractVerifier {
+export abstract class AbstractAbiProvider {
   constructor(public apiUrl: string, public explorerUrl: string) {}
 
   abstract verify(submittedContract: SubmittedContract, compilerAbstract: CompilerAbstract): Promise<VerificationResponse>
