+++ apps/contract-interaction/src/app/views/LookupView.tsx
@@ -5,7 +5,7 @@ import type { LookupResponse, VerifierIdentifier } from '../types'
 import { VERIFIERS } from '../types'
 import { AppContext } from '../AppContext'
 import { CustomTooltip } from '@remix-ui/helper'
-import { getVerifier } from '../Verifiers'
+import { getVerifier } from '../abiProviders'
 import { useNavigate } from 'react-router-dom'
 import { VerifyFormContext } from '../VerifyFormContext'
 import { useSourcifySupported } from '../hooks/useSourcifySupported'
